modify vps-create-from-tar.sh

+++ tools/vps-create-from-tar.sh
@@ -6,11 +6,11 @@ XENCONF=/etc/xen
 XENLOG=/var/log/xen
 XENTOOLLOG=/var/log/xen-tools
 
-NAME="vps20"
+NAME="vps21"
 VCPU="1"
 VRAM="2048"
-SWAP="1024"
-DISK="5G"
+SWAP="2048"
+DISK="50G"
 
 OS="ubuntu"
 #OS="centos5"
@@ -18,9 +18,9 @@ OS="ubuntu"
 #OS="debian"
 #OS="gentoo"
 #OS="arch"
-#IPADDRESS="119.254.32.170"
+#ADDRESS="119.254.32.170"
 #GATEWAY="119.254.32.161"
-IPADDRESS="119.254.35.107"
+ADDRESS="119.254.35.107"
 GATEWAY="119.254.35.97"
 NETMASK="255.255.254.0"
 
@@ -38,7 +38,7 @@ vcpus = "$VCPU"
 maxmem = "$VRAM"
 memory = "$VRAM"
 name = "$NAME"
-vif = [ "vifname=$NAME,mac=`genXenMac`,ip=$IPADDRESS,rate=500KB/s,bridge=xenbr0" ]
+vif = [ "vifname=$NAME,mac=`genXenMac`,ip=$ADDRESS,rate=500KB/s,bridge=xenbr0" ]
 disk = [ "file:$XENHOME/$NAME.img,xvda1,w","file:$XENSWAP/$NAME.swp,xvdb1,w" ]
 #root = "/dev/sda1"
 #extra = "fastboot"
@@ -54,7 +54,7 @@ vcpus = "$VCPU"
 maxmem = "$VRAM"
 memory = "$VRAM"
 name = "$NAME"
-vif = [ "vifname=$NAME,mac=`genXenMac`,ip=$IPADDRESS,bridge=xenbr0" ]
+vif = [ "vifname=$NAME,mac=`genXenMac`,ip=$ADDRESS,bridge=xenbr0" ]
 disk = [ "file:$XENHOME/$NAME.img,sda1,w","file:$XENSWAP/$NAME.swp,sda2,w" ]
 root = "/dev/sda1"
 extra = "fastboot"
@@ -65,14 +65,23 @@ on_crash = "restart"
 __END__
 fi
 
-dd if=/dev/zero of=$XENHOME/$NAME.img bs=1 count=1 seek=$DISK
-mkfs.ext3 $XENHOME/$NAME.img
-mount -o loop $XENHOME/$NAME.img /mnt
-tar -zxSf $XENHOME/$OS.tar.gz -C /mnt/
-umount /mnt
+if [ ! -f $XENHOME/$NAME.img ];
+then
+	dd if=/dev/zero of=$XENHOME/$NAME.img bs=1 count=1 seek=$DISK
+	mkfs.ext3 $XENHOME/$NAME.img
+	mount -o loop $XENHOME/$NAME.img /mnt
+	tar -zxSf $XENHOME/$OS.tar.gz -C /mnt/
+	umount /mnt
+else
+	echo $XENHOME/$NAME.img exists!;
+	exit 1;
+fi
 
-dd if=/dev/zero of=$XENSWAP/$NAME.swp bs=1024 count=`expr $SWAP \* 1024`
-mkswap $XENSWAP/$NAME.swp
+if [ ! -f $XENHOME/$NAME.swp ];
+then
+	dd if=/dev/zero of=$XENSWAP/$NAME.swp bs=1024 count=`expr $SWAP \* 1024`
+	mkswap $XENSWAP/$NAME.swp
+fi
 
 mount -o loop $XENHOME/$NAME.img /mnt
 
@@ -84,7 +93,7 @@ iface lo inet loopback
 
 auto eth0
 iface eth0 inet static
-address $IPADDRESS
+address $ADDRESS
 gateway $GATEWAY
 netmask $NETMASK
 __END__
@@ -95,7 +104,7 @@ DEVICE=eth0
 BOOTPROTO=none
 ONBOOT=yes
 TYPE=Ethernet
-IPADDR=$IPADDRESS
+IPADDR=$ADDRESS
 GATEWAY=$GATEWAY
 NETMASK=$NETMASK
 __END__
@@ -106,7 +115,7 @@ DEVICE="eth0"
 BOOTPROTO="static"
 DNS1="8.8.8.8"
 GATEWAY="$GATEWAY"
-IPADDR="$IPADDRESS"
+IPADDR="$ADDRESS"
 IPV6INIT="no"
 MTU="1500"
 NETMASK="$NETMASK"
@@ -117,13 +126,13 @@ __END__
 elif [ $OS == "gentoo" ]
 then
 cat >/mnt/etc/conf.d/net <<-__END__
-config_eth0="$IPADDRESS netmask $NETMASK"
+config_eth0="$ADDRESS netmask $NETMASK"
 routes_eth0="default via $GATEWAY"
 __END__
 elif [ $OS == "arch" ]
 then
 cat >/mnt/etc/rc.conf <<-__END__
-config_eth0="$IPADDRESS netmask $NETMASK"
+config_eth0="$ADDRESS netmask $NETMASK"
 routes_eth0="default via $GATEWAY"
 
 LOCALE="en_US.utf8"
@@ -139,7 +148,7 @@ MODULES=()
 HOSTNAME="arch"
 USELVM="no"
 INTERFACES=(eth0)
-eth0="eth0 $IPADDRESS netmask $NETMASK"
+eth0="eth0 $ADDRESS netmask $NETMASK"
 gateway="default gw $GATEWAY"
 ROUTES=(gateway)
 DAEMONS=(syslog-ng network crond sshd)
